+++ app/src/main/java/ar/edu/utn/frsf/dam/isi/laboratorio02/dao/PedidoDetalleDao.java
@@ -16,7 +16,7 @@ public interface PedidoDetalleDao {
     List<PedidoDetalle> getAll();
 
     @Insert
-    int insert(PedidoDetalle pd);
+    long insert(PedidoDetalle pd);
 
     @Delete
     void delete(PedidoDetalle pd);
